+++ src/DGtal/geometry/surfaces/estimation/CDigitalSurfaceLocalEstimator.h
@@ -46,8 +46,9 @@
 #include "DGtal/geometry/surfaces/estimation/CSurfelLocalEstimator.h"
 //////////////////////////////////////////////////////////////////////////////
 
-namespace DGtal
-{
+// @since 0.8 In DGtal::concepts
+namespace DGtal {
+  namespace concepts {
 
   /////////////////////////////////////////////////////////////////////////////
   // class CDigitalSurfaceLocalEstimator
@@ -107,7 +108,7 @@ namespace DGtal
 
     typedef typename T::Surface Surface;
     typedef typename Surface::DigitalSurfaceContainer DigitalSurfaceContainer;
-    BOOST_CONCEPT_ASSERT(( CDigitalSurfaceContainer< DigitalSurfaceContainer > ));
+    BOOST_CONCEPT_ASSERT(( concepts::CDigitalSurfaceContainer< DigitalSurfaceContainer > ));
     typedef typename T::Surfel Surfel;
     BOOST_STATIC_ASSERT(( ConceptUtils::SameType< Surfel, typename Surface::Surfel >::value ));
     BOOST_CONCEPT_USAGE( CDigitalSurfaceLocalEstimator )
@@ -125,6 +126,7 @@ namespace DGtal
 
   }; // end of concept CDigitalSurfaceLocalEstimator
 
+} // namespace concepts
 } // namespace DGtal
 
 //                                                                           //